Move FileId, FileSize, MimeType in file package
